added authorization in account force delete

+++ app/Http/Controllers/AccountController.php
@@ -70,6 +70,8 @@ class AccountController extends Controller
     {
         $account = Account::withTrashed()->find($id);
 
+        Gate::authorize('forceDelete', $account);
+
         $account->forceDelete();
 
         return response()->json(['message' => 'Account force deleted successfully'], 204);
